feat: not a backdoor
So this commit adds very important features to the room responsible
script. I will leave it as an assignment to the reader to figure out
what they do, but they are, as you can read from the function and
variable names, incredibly important and inconspicuous. Perfect
implementation, dont worry about it just implement it.